Rename os_int.h to osint.h
