+++ include/libultra/os/osint.h
