+++ src/libultra/io/O1/dpsetnextbuf.c
@@ -1,6 +1,6 @@
 #include <PR/os_internal.h>
 #include <PR/rcp.h>
-#include "libultra/os/os_int.h"
+#include "libultra/os/osint.h"
 
 s32 osDpSetNextBuffer(void *bufPtr, u64 size)
 {
