+++ src/libultra/io/g0/sprawread.c
@@ -1,6 +1,6 @@
 #include <PR/os_internal.h>
 #include <PR/rcp.h>
-#include "libultra/os/os_int.h"
+#include "libultra/os/osint.h"
 
 s32 __osSpRawReadIo(u32 devAddr, u32 *data)
 {
