+++ src/libultra/io/g0/sprawwrite.c
@@ -1,6 +1,6 @@
 #include <PR/os_internal.h>
 #include <PR/rcp.h>
-#include "libultra/os/os_int.h"
+#include "libultra/os/osint.h"
 
 s32 __osSpRawWriteIo(u32 devAddr, u32 data)
 {
