+++ src/libultra/os/O1/settime.c
@@ -1,5 +1,5 @@
 #include <PR/os_internal.h>
-#include "libultra/os/os_int.h"
+#include "libultra/os/osint.h"
 
 void osSetTime(OSTime time){
     __osCurrentTime = time;
