+++ src/libultra/os/O1/yieldthread.c
@@ -1,5 +1,5 @@
 #include <PR/os_internal.h>
-#include "libultra/os/os_int.h"
+#include "libultra/os/osint.h"
 
 void osYieldThread(void){
     register u32 saveMask = __osDisableInt();
